move lxc wrapper build
It turns out that anything using go-embed in a library (even if that
particular package is not used) will be required if someone uses a package
in a library. For example, since we had github.com/anuvu/stacker requiring
a lxc-wrapper directory, something that imported
github.com/anuvu/stacker/oci needed to also have this directory.

Let's move this lxc-wrapper gunk to the main package, so that users don't
have to build anything like this.

Users of the API will have to populate the stacker config's new EmbeddedFS
member in order to actually spawn containers, but we can leave that as an
exercise for users for now.

This at least enables library users of stacker's other interesting packages
(oci, mount, etc.) to not have to use this.
