--- lxc-wrapper.go
@@ -1,6 +1,0 @@
-package stacker
-
-import "embed"
-
-//go:embed lxc-wrapper/lxc-wrapper
-var embeddedFS embed.FS
